Relaciones solucionadas, y no deja borrar entidades implicadas en la relacion, asimismo se soluciono la curp de trabajadores

+++ app/Http/Controllers/BaseUserController.php
@@ -2,9 +2,11 @@
 
 namespace App\Http\Controllers;
 
+use MongoDB\BSON\ObjectId;
 use App\Models\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Hash;
+use App\Models\Asignacion;
 
 abstract class BaseUserController extends Controller
 {
@@ -136,8 +138,9 @@ abstract class BaseUserController extends Controller
         ];
 
         if (in_array($this->role, ['trabajador', 'administrador'])) {
-            $rules['curp '] = 'required|string|max:18|unique:users,curp,' . $id;
+            $rules['curp'] = 'required|string|max:18|unique:users,curp,' . $id;
             $rules['rfc']  = 'required|string|max:13|unique:users,rfc,' . $id;
+
         }
 
         $validated = $request->validate($rules);
@@ -189,9 +192,20 @@ abstract class BaseUserController extends Controller
             return response()->json(['success' => false, 'message' => ucfirst($this->role) . ' no encontrado'], 404);
         }
 
+        $userIdObject = new ObjectId($user->getKey());
+
+        // Comprobar si el usuario tiene asignaciones activas
+        if (Asignacion::where('idUsuario', $userIdObject)->exists()) {
+            return response()->json([
+                'success' => false,
+                'message' => 'No se puede eliminar el trabajador porque tiene asignaciones activas.'
+            ], 400);
+        }
+
         $user->delete();
 
         return response()->json(['success' => true, 'message' => ucfirst($this->role) . ' eliminado correctamente']);
     }
+
 }
